--- plugin-server/src/worker/ingestion/event-pipeline/createEventStep.ts
@@ -1,11 +1,0 @@
-import { Person, PreIngestionEvent, RawKafkaEvent } from '../../../types'
-import { EventPipelineRunner } from './runner'
-
-export function createEventStep(
-    runner: EventPipelineRunner,
-    event: PreIngestionEvent,
-    person: Person,
-    processPerson: boolean
-): RawKafkaEvent {
-    return runner.eventsProcessor.createEvent(event, person, processPerson)
-}
