--- plugin-server/src/worker/ingestion/event-pipeline/emitEventStep.ts
@@ -1,6 +1,0 @@
-import { RawKafkaEvent } from '../../../types'
-import { EventPipelineRunner } from './runner'
-
-export function emitEventStep(runner: EventPipelineRunner, event: RawKafkaEvent): [Promise<void>] {
-    return [runner.eventsProcessor.emitEvent(event)]
-}
