--- plugin-server/src/worker/ingestion/event-pipeline/processPersonsStep.ts
@@ -1,26 +1,0 @@
-import { PluginEvent } from '@posthog/plugin-scaffold'
-import { DateTime } from 'luxon'
-
-import { Person, Team } from '~/src/types'
-
-import { PersonState } from '../person-state'
-import { EventPipelineRunner } from './runner'
-
-export async function processPersonsStep(
-    runner: EventPipelineRunner,
-    event: PluginEvent,
-    team: Team,
-    timestamp: DateTime,
-    processPerson: boolean
-): Promise<[PluginEvent, Person, Promise<void>]> {
-    const [person, kafkaAck] = await new PersonState(
-        event,
-        team,
-        String(event.distinct_id),
-        timestamp,
-        processPerson,
-        runner.hub.db
-    ).update()
-
-    return [event, person, kafkaAck]
-}
