--- plugin-server/src/worker/ingestion/event-pipeline/transformEventStep.ts
@@ -1,18 +1,0 @@
-import { PluginEvent } from '@posthog/plugin-scaffold'
-
-import { HogTransformerService, TransformationResult } from '../../../cdp/hog-transformations/hog-transformer.service'
-import { droppedEventCounter } from './metrics'
-
-export async function transformEventStep(
-    event: PluginEvent,
-    hogTransformer: HogTransformerService | null
-): Promise<TransformationResult> {
-    if (!hogTransformer) {
-        return { event, invocationResults: [], messagePromises: [] }
-    }
-    const result = await hogTransformer.transformEventAndProduceMessages(event)
-    if (!result.event) {
-        droppedEventCounter.inc()
-    }
-    return result
-}
