+++ plugin-server/tests/main/ingestion-queues/run-async-handlers-event-pipeline.test.ts
@@ -20,13 +20,13 @@ import Redis from 'ioredis'
 import LibrdKafkaError from 'node-rdkafka/lib/error'
 
 import { KAFKA_EVENTS_JSON } from '../../../src/config/kafka-topics'
+import { processOnEventStep } from '../../../src/main/ingestion-queues/batch-processing/runAsyncHandlersStep'
 import { buildOnEventIngestionConsumer } from '../../../src/main/ingestion-queues/on-event-handler-consumer'
 import { Hub, ISOTimestamp } from '../../../src/types'
 import { DependencyUnavailableError } from '../../../src/utils/db/error'
 import { closeHub, createHub } from '../../../src/utils/db/hub'
 import { PostgresUse } from '../../../src/utils/db/postgres'
 import { UUIDT } from '../../../src/utils/utils'
-import { processOnEventStep } from '../../../src/worker/ingestion/event-pipeline/runAsyncHandlersStep'
 import { setupPlugins } from '../../../src/worker/plugins/setup'
 import { teardownPlugins } from '../../../src/worker/plugins/teardown'
 import {
